moved `qml.device` to `qml.devices` module
